Topology Change
Channelized stepper drivers and min/max enstops, 6 channels total for
axis. Going to include 6 12v switchable mosfets and 3 thermister
sockets. While the goal of the board is to be a general purpose TinyG
platform the simple inclusion of thermister circuitry will  lend itself
to the popularity of open sourcer 3D printing.